implementation of backup and restore for telegram
