+++ .gitignore
@@ -26,3 +26,4 @@ potato/*
 src/lib/db/poracle.sqlite
 .geoCache
 /.cache/cache/
+backups/*
